pyproject.toml
[tool.poetry]
name = "workflows"
version = "0.1.0"
description = "workflows to ease life"
authors = ["Jooeon Kim"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pypdf = "^5.3.0"  # unlock_pdf.py
pillow = "^11.2.0"  # merge_images.py


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
